+++ CHANGES.md
@@ -13,10 +13,14 @@ C++ Changes:
 * Major rework of the C++ Solver backend
   * The `CySolverResult` storage structure is now the owner of: the solution data, the dense outputs, _and_ the solver itself (before the solver was stand alone). This ensures that the solver is alive for the dense outputs which, depending on settings, may need to make calls to the solver even after the integration is complete. 
   * Removed many pointers in favor of direct access of variables. This was due to some variables moving and hanging pointers causing crashes.
+  * Drastically reduced the size of `CySolverDense` and moved its state data to a heap allocated array.
+* `num_y` and `num_extra` are now size_t type instead of unsigned ints.
+* `integration method` is not int instead of unsigned int.
 
 Fixes:
-* Fixed issue where if t_eval was less than t_end it could cause an access violation
+* Fixed issue where if t_eval was less than t_end it could cause an access violation.
 * Fixed issue where dense output was sometimes being created twice when t_eval was provided.
+* Addressed several compile warnings.
 
 #### v0.11.2 (2024-11-12)